refactor(mcp): move shared MCP infra out of the chat tool folder
tools/tool_implementations/mcp/ keeps only the chat Tool implementation;
everything shared between chat, the admin API, and Craft moves to the
MCP feature package:

- mcp_client.py -> server/features/mcp/client.py (protocol/session
  plumbing; sync + async wrappers)
- mcp_ssrf.py -> server/features/mcp/ssrf.py
- OnyxTokenStorage, make_oauth_provider, MCPOauthState, the redis key
  helpers, and the token-expiry/refresh-preservation helpers move out of
  the 2.7k-line router module into server/features/mcp/oauth.py; api.py
  keeps routes and route-only flow helpers

This also deletes the deferred function-level imports in MCPTool.run
that existed only because the OAuth machinery lived inside api.py.
Pure moves + import updates; no logic changes.
