+++ backend/tests/unit/onyx/server/features/mcp/test_mcp_ssrf.py
@@ -16,10 +16,10 @@ from onyx.auth.oauth_token_manager import OAuthFlowParams
 from onyx.error_handling.error_codes import OnyxErrorCode
 from onyx.error_handling.exceptions import OnyxError
 from onyx.server.features.mcp import api
+from onyx.server.features.mcp import ssrf as mcp_ssrf
 from onyx.server.security.models import SecuritySettings
 from onyx.server.security.models import SSRFProtectionLevel
 from onyx.server.security.store import _build_env_defaults
-from onyx.tools.tool_implementations.mcp import mcp_ssrf
 from onyx.utils.url import SSRFException
 
 # IP literals (and the named host localhost, blocked by name) so validation never
